Blade Throw Buff
Blade throws that hit napes have a 1.5x multiplier to damage.

Design philosophy for buff:
Blade throws can be hard to hit and take away an entire blade set. Even with the multiplier it is not a good idea to throw blades willy nilly, but now you get rewarded if you do make contact.